Refactoring and parser now reports errors at the end
