--- nolang/util/log.py
@@ -1,5 +1,0 @@
-
-from sys import stderr
-
-def log_error(message: str):
-    print(f'nolang: {message}', file=stderr)
