+++ supplemental/parse_errors.nl
@@ -1,0 +1,13 @@
+# Unexpected token
+if
+==
++
+
+# The following causes some cascading exceptions, this should be fixed!
+
+    nolout('hi')
+
+if True
+    nolout('balls')
+x
+8 = 9
